requirements.txt
django
djangorestframework
gunicorn
psycopg2-binary
python-dotenv
dj-database-url
whitenoise
